Analyzer MoveFunctionsOutOfAnyPass refactoring
